Updating tests.

+++ projects/ngkit/src/lib/services/storage/session.ts
@@ -8,17 +8,13 @@ export class SessionStorage implements StorageDriver {
   /**
    * Create a new instance of the service.
    */
-  constructor(private config: Config) {
-
-  }
+  constructor(private config: Config) {}
 
   /**
    * Get item from session storage.
    */
   async get(key: string): Promise<any> {
-    await this.load;
-
-    if (window?.sessionStorage) {
+    if (window !== undefined && window.sessionStorage) {
       return await sessionStorage.getItem(key);
     }
   }
@@ -27,7 +23,7 @@ export class SessionStorage implements StorageDriver {
    * Set an item to session storage.
    */
   async set(key: string, value: any): Promise<any> {
-    if (window?.sessionStorage) {
+    if (window !== undefined && window.sessionStorage) {
       return await sessionStorage.setItem(key, value);
     }
   }
@@ -36,7 +32,7 @@ export class SessionStorage implements StorageDriver {
    * Remove an item from session storage.
    */
   async remove(key: string): Promise<void> {
-    if (window?.sessionStorage) {
+    if (window !== undefined && window.sessionStorage) {
       return await sessionStorage.removeItem(key);
     }
   }
@@ -45,7 +41,7 @@ export class SessionStorage implements StorageDriver {
    * Clear session storage.
    */
   async clear(): Promise<void> {
-    if (window?.sessionStorage) {
+    if (window !== undefined && window.sessionStorage) {
       return await sessionStorage.clear();
     }
   }
